+++ src/modules/auth/hooks/useRequestConfirmationMail.ts
@@ -2,15 +2,13 @@
 
 import { VerificationType } from '@w3block/sdk-id';
 
-import { useMutation } from '@tanstack/react-query';
+import { UseMutationResult, useMutation } from '@tanstack/react-query';
 import { useGetW3blockIdSDK } from '../../shared/hooks/useGetW3blockIdSDK';
 import { useCompanyConfig } from '../../shared/hooks/useCompanyConfig';
 import { PixwayAPIRoutes } from '../../shared/enums/PixwayAPIRoutes';
-import { removeDoubleSlashesOnUrl } from '../../shared/utils/removeDuplicateSlahes';
 import { PixwayAppRoutes } from '../../shared/enums/PixwayAppRoutes';
+import { useResolveCallbackUrl } from '../../shared/hooks/useResolveCallbackUrl';
 import { handleNetworkException } from '../../shared/utils/handleNetworkException';
-
-
 
 interface Payload {
   email: string;
@@ -19,9 +17,14 @@ interface Payload {
   verificationType?: 'numeric' | 'invisible';
 }
 
-export const useRequestConfirmationMail = (): any => {
+export const useRequestConfirmationMail = (): UseMutationResult<
+  unknown,
+  unknown,
+  Payload
+> => {
   const getSDK = useGetW3blockIdSDK();
-  const { companyId, appBaseUrl, connectProxyPass } = useCompanyConfig();
+  const { companyId } = useCompanyConfig();
+  const { resolveCallbackUrl } = useResolveCallbackUrl();
 
   return useMutation(
     [PixwayAPIRoutes.REQUEST_CONFIRMATION_MAIL],
@@ -34,10 +37,7 @@ export const useRequestConfirmationMail = (): any => {
             (verificationType as VerificationType) ??
             VerificationType.Invisible,
           tenantId: tenantId ?? companyId,
-          callbackUrl: removeDoubleSlashesOnUrl(
-            callbackPath ??
-              appBaseUrl + connectProxyPass + PixwayAppRoutes.COMPLETE_SIGNUP
-          ).toString() ?? '/',
+          callbackUrl: resolveCallbackUrl(PixwayAppRoutes.COMPLETE_SIGNUP, callbackPath),
         });
       } catch (error) {
         console.error('Erro ao solicitar e-mail de confirmação:', error);